requirements.txt
numpy
matplotlib
opencv-python
open3d-cpu